+++ changelog.md
@@ -10,6 +10,7 @@
 ## Bug fixes
 
 - Fix reporting error based on http responses from the Aura-API with an invalid JSON body. Earlier the client would report JSONDecodeError instead of showing the actual issue.
+- Fixed a bug where retryable queries wouldnt work with bookmarks.
 
 ## Improvements